+++ webapp/src/pages/EditChatroomDetailsPage/EditChatroomDetailsPage.tsx
@@ -24,7 +24,7 @@ import NotFoundPage from '../NotFoundPage/NotFoundPage';
 import './EditChatroomDetailsPage.css';
 import { useAuth } from '../../shared/hooks/UseAuth';
 import { useGetChatroomMember } from '../../shared/hooks/UseGetChatroomMember';
-import { handleRequestError } from '../../shared/utils';
+import { handleRequestError } from '../../shared/utils/HandleRequestError';
 
 export default function EditChatroomDetailsPage() {
   const { pathname } = useLocation();
