+++ webapp/src/pages/JoinChatroomPage/JoinChatroomPage.tsx
@@ -14,7 +14,7 @@ import { useData } from '../../shared/hooks/UseData';
 import { useNavigation } from '../../shared/hooks/UseNavigation';
 import { chatApi } from '../../shared/services/ApiService';
 import { AVATAR_EP_URL, CHATROOM_URL } from '../../shared/urls';
-import { handleRequestError } from '../../shared/utils';
+import { handleRequestError } from '../../shared/utils/HandleRequestError';
 
 export default function JoinChatroomPage() {
   const { chatroomId } = useParams();
